--- condition-checker/src/test/java/com/mersey/rowing/club/condition_checker/ConditionCheckerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mersey.rowing.club.condition_checker;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConditionCheckerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
