+++ condition-checker/src/test/java/com/mersey/rowing/club/condition_checker/applicationTests/controller/boats/BoatCapabilityClientTests.java
@@ -1,8 +1,8 @@
-package com.mersey.rowing.club.condition_checker.controller.boats;
+package com.mersey.rowing.club.condition_checker.applicationTests.controller.boats;
 
-import com.mersey.rowing.club.condition_checker.mock.MockOpenWeatherResponseGenerator;
+import com.mersey.rowing.club.condition_checker.controller.boats.BoatCapabilityClient;
+import com.mersey.rowing.club.condition_checker.mockSetup.MockOpenWeatherResponseGenerator;
 import com.mersey.rowing.club.condition_checker.model.response.BoatsAllowed;
-import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -11,10 +11,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import static com.mersey.rowing.club.condition_checker.mock.MockOpenWeatherResponseGenerator.*;
+import static com.mersey.rowing.club.condition_checker.mockSetup.MockOpenWeatherResponseGenerator.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@SpringBootTest
+@SpringBootTest(properties = { "open-weather-api.key=test","open-weather-api.baseUrl=http://localhost:5050" })
 public class BoatCapabilityClientTests {
 
     @Autowired
