+++ condition-checker/src/test/java/com/mersey/rowing/club/condition_checker/applicationTests/controller/openWeather/OpenWeatherApiClientTests.java
@@ -1,7 +1,7 @@
-package com.mersey.rowing.club.condition_checker.controller.openWeather;
+package com.mersey.rowing.club.condition_checker.applicationTests.controller.openWeather;
 
 
-import com.mersey.rowing.club.condition_checker.WiremockBaseTests;
+import com.mersey.rowing.club.condition_checker.applicationTests.WiremockBaseTests;
 import com.mersey.rowing.club.condition_checker.controller.openweather.OpenWeatherApiClient;
 import com.mersey.rowing.club.condition_checker.model.openweatherapi.OpenWeatherResponse;
 import org.junit.jupiter.api.Test;
